Tuesday stream progress
We finished off a lot more with the swapchain and are almost in a
position to present it, but something doesn't seem to be quite right.
We've got a render pass setup and ready to go, along with a recording
into a command buffer that should give us a blue screen.

Our blue screen of progress is close!